Move id_serialization away
